+++ changes.md
@@ -1,5 +1,8 @@
 # Changes and Feature List
 
+## Version 1.4.1 Support for sending arbritrary messages
+* Added Service for sending arbritrary ESP2 messages
+
 ## Version 1.4.0 ESP3 Support (USB300)
 * Docs about gateway usage added.
 * Added EEPs F6-02-01 and F6-02-02 as sender EEP for lights so that regular switch commands can be sent from Home Assistant.